Back-End/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
Django==2.1.1
django-estimators==0.2.1
djangorestframework==3.11.0
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
Pillow==7.2.0
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
sqlparse==0.3.1
urllib3==1.25.10
